--- policy_learning_algorithm.py
@@ -1,21 +1,0 @@
-"""
-Whatever policy learning algorithm we use.
-"""
-
-class PolicyLearningAlgorithm:
-    
-    # Chooses the optimal action given the state
-    def optimal_action(self, state):
-        return None
-    
-    # Initializes the algorithm
-    def initialize(self):
-        return
-    
-    # Saves the policy learned so far in some form
-    def save(self):
-        return
-    
-    # Loads the policy
-    def load(self):
-        return
